% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_name.r
\name{sw_name}
\alias{sw_name}
\title{sw_name}
\usage{
sw_name()
}
\description{
Pasta-based Star Wars name generator.
}
\details{
No, it's exactly what it sounds like.
}
\examples{
pastarwars::sw_name()

}
\references{
\url{https://twitter.com/ollymoss/status/943254964436197378}
}

